# specs/001-reimbursement-screen/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Elosaúde Reimbursement API
  description: API endpoints for reimbursement management in the Elosaúde platform
  version: 1.0.0
  contact:
    name: Elosaúde Development Team
servers:
  - url: http://localhost:8000/api
    description: Development server
  - url: https://api.elosaude.com.br/api
    description: Production server
security:
  - bearerAuth: []
...
paths:
...
  /reimbursements/documents/:
    get:
      tags: [Documents]
      summary: List reimbursement documents
      parameters:
        - name: reimbursement
          in: query
          schema:
            type: integer
          description: Filter by reimbursement ID
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReimbursementDocument'

    post:
      tags: [Documents]
      summary: Upload a document
      description: Uploads a document file. Returns document ID for association with reimbursement.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - document_type
              properties:
                file:
                  type: string
                  format: binary
                  description: Document file (PDF, JPG, PNG, max 10MB)
                document_type:
                  type: string
                  enum: [INVOICE, PRESCRIPTION, REPORT, RECEIPT, OTHER]
                description:
                  type: string
                  maxLength: 200
                reimbursement:
                  type: integer
                  description: Optional reimbursement ID to associate immediately
      responses:
        '201':
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReimbursementDocument'
        '400':
          $ref: '#/components/responses/ValidationError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ReimbursementDocument:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        reimbursement:
          type: integer
        document_type:
          type: string
          enum: [INVOICE, PRESCRIPTION, REPORT, RECEIPT, OTHER]
        document_type_display:
          type: string
          readOnly: true
        description:
          type: string
          maxLength: 200
        file:
          type: string
          format: uri
          writeOnly: true
        file_url:
          type: string
          format: uri
          readOnly: true
        uploaded_at:
          type: string
          format: date-time
          readOnly: true
...
  responses:
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            additionalProperties:
              type: array
              items:
                type: string
          example:
            requested_amount: ["Este campo é obrigatório."]
            service_date: ["A data do serviço não pode ser futura."]